move search to tracker app
